requirements.txt
aiogram==2.11.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==41.0.5
defusedxml==0.7.1
ecdsa==0.19.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
jira==3.5.1
magic-filter==1.0.12
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
Pillow==10.0.0
propcache==0.2.0
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2024.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.23
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
